perf(gpu): move subgrid packing to GPU
Instead of packing and unpacking the subgrids on the CPU every timestep
we now create a connectivity graph of the subgrids on the first time
step then use that to index directly to the octs and build the subgrid
in shared memory. This entirely eliminates the `subgrids_host` and
`subgrids_device` arrays, saving lots of memory, and improves integrator
performance by ~115x at a grid size of 256^3 cells.

Notes:
- CUDA Fortran assumes that all allocatable arrays use 1-based indexing,
  even if you set the indexing to something else. This can result in
  segfaults and silent errors if you try to use an allocatable array
  with 0-based indexing
- The GPU arrays are all allocated as C-style arrays then a Fortran
  pointer is created to point at them. This eliminates some superfluous
  cudaMallocs and cudaFrees that CUDA Fortran does under the hood which
  can cause 2-10x slow downs of some kernels. Using deferred size arrays
  on the GPU is also required to get this performance gain.